DRY authorised events and harden validations
