+++ app/models/generic_event/move_lockout.rb
@@ -1,6 +1,7 @@
 class GenericEvent
   class MoveLockout < GenericEvent
     include MoveEventValidations
+    include AuthoriserValidations
 
     enum reason: {
       unachievable_ptr_request: 'unachievable_ptr_request', # (PECS - policy only)
@@ -16,28 +17,6 @@ class GenericEvent
 
     validates :from_location_id, presence: true
     validates :reason, inclusion: { in: reasons }, if: -> { reason }
-
-    validates_each :authorised_at, if: -> { authorised_at.present? } do |record, attr, value|
-      Time.zone.iso8601(value)
-    rescue ArgumentError
-      record.errors.add(attr, 'must be formatted as a valid ISO-8601 date-time')
-    end
-
-    def authorised_at=(authorised_at)
-      details['authorised_at'] = authorised_at
-    end
-
-    def authorised_at
-      details['authorised_at']
-    end
-
-    def authorised_by=(authorised_by)
-      details['authorised_by'] = authorised_by
-    end
-
-    def authorised_by
-      details['authorised_by']
-    end
 
     def reason=(reason)
       details['reason'] = reason
