+++ spec/factories/generic_event.rb
@@ -51,7 +51,7 @@ FactoryBot.define do
         from_location_id: create(:location).id,
         reason: 'no_space',
         authorised_at: Time.zone.now.iso8601,
-        authorised_by: Faker::Name.name,
+        authorised_by: 'PMU',
       }
     end
   end
@@ -102,7 +102,7 @@ FactoryBot.define do
     details do
       {
         authorised_at: Time.zone.now.iso8601,
-        authorised_by: Faker::Name.name,
+        authorised_by: 'PMU',
       }
     end
   end
@@ -112,7 +112,7 @@ FactoryBot.define do
     details do
       {
         authorised_at: Time.zone.now.iso8601,
-        authorised_by: Faker::Name.name,
+        authorised_by: 'PMU',
       }
     end
   end
