+++ spec/models/generic_event/move_operation_tornado_spec.rb
@@ -2,16 +2,5 @@ RSpec.describe GenericEvent::MoveOperationTornado do
   subject(:generic_event) { build(:event_move_operation_tornado) }
 
   it_behaves_like 'a move event'
-
-  context 'when authorised_at is supplied' do
-    it 'is valid when the authorised_at value is a valid iso8601 datetime' do
-      generic_event.authorised_at = '2020-06-16T10:20:30+01:00'
-      expect(generic_event).to be_valid
-    end
-
-    it 'is invalid when the authorised_at value is not a valid iso8601 datetime' do
-      generic_event.authorised_at = '16-06-2020 10:20:30+01:00'
-      expect(generic_event).not_to be_valid
-    end
-  end
+  it_behaves_like 'an authorised event'
 end
